+++ src/encoding/mod.rs
@@ -2,6 +2,6 @@ mod errors;
 #[cfg(feature = "evm")]
 mod evm;
 mod models;
-mod router_encoder;
 mod strategy_encoder;
 mod swap_encoder;
+mod tycho_encoder;